# GAL: Send all buffer assignments at once rather than individually