The GAL Multithreading layer is good at reducing the cost of calls that would take longer on the backend, but the downside is an extra cost for loading commands into the command queue, and resource references/spans into circular queues.

The `(int first, BufferRange[] ranges)` method call has very significant performance implications when the bindings are spread out, which they generally always are in Vulkan, as the method ends up being called a lot of times every buffer update. This change makes it so that these methods are only called a maximum of one time per draw.

Significantly improves GPU thread performance in SMO, Pokemon Scarlet/Violet, and any other games that tend to bind many constant buffers at once. Probably not as much in OpenGL.

This change is already in LDN 3.0.