@riperiperi 2022-11-21
Super Mario Odyssey (best case FPS)

Before:
![image](https://user-images.githubusercontent.com/6294155/203180131-abfe0a4a-c6cb-485a-81d5-24dfc7b898df.png)

After:
![image](https://user-images.githubusercontent.com/6294155/203180429-bb485031-c7fd-45fa-9c20-c169f479bf9d.png)

Maybe I should get actual metrics.