@Littl3Guy 2022-11-22
Fixes the visual glitch in Mario Golf: Super Rush
Before:
![mg](https://user-images.githubusercontent.com/115218507/203252716-e5dcaed8-541e-4786-8882-4bbd80d1d9a0.png)

After:
![mg2](https://user-images.githubusercontent.com/115218507/203252836-db19c8c1-0c2e-4c81-bcab-757242c426b9.png)