@lostromb 2022-11-24
Breath of the Wild seems to regress about 7% with this change, from 39 down to 36 FPS.
I'm still trying to get more accurate numbers across different games...