@gdkchan 2022-11-24
> Breath of the Wild seems to regress about 7% with this change, from 39 down to 36 FPS. I'm still trying to get more accurate numbers across different games...

Where exactly in the game are you testing this? SInce performance may vary a lot depending on the area and view.