@lostromb 2022-11-24
Just on the first vista looking at the horizon from the plateau. But right now I'm also checking Mario Odyssey to validate that things run faster over there like we assume. And other games too if I have time tonight

![image](https://user-images.githubusercontent.com/918053/203677609-e0a11b27-ecb2-4539-9e5a-c99511932a34.png)