@lostromb 2022-11-24
Actual numbers for different games. And I ran the BOTW numbers again since I was right to doubt them the first time.

Mario Odyssey - **66.9 -> 77.0 FPS (15% faster)**
Capturing the pole overlooking Metro Kingdom
![image](https://user-images.githubusercontent.com/918053/203681833-216b9462-d86c-4ba5-b824-1410ea3abb2b.png)

Smash Ultimate - **42.9 -> 50.9 FPS (18% faster)**
8 CPU Marios in Fountain of Dreams
![image](https://user-images.githubusercontent.com/918053/203682119-f4ff669d-d01b-4ced-9a26-12c14859ee9c.png)

Breath of the Wild - **39.3 -> 40.6 FPS** (3% faster)
Great plateau, first vista
![image](https://user-images.githubusercontent.com/918053/203683810-7d697e45-dff4-4fb6-a995-896795ef059f.png)

Pokemon Arceus - **36.0 -> 35.5 FPS (negligible decrease)**
Standing on the bridge of town looking at HQ
![image](https://user-images.githubusercontent.com/918053/203682334-bed75f16-15f3-4718-9007-a43efad83261.png)